moved the code for the task
