+++ .gitignore
@@ -17,3 +17,5 @@ build/
 
 bundle.js
 bundle.js.map
+
+temp-*
