+++ src/task/task-constants.js
@@ -1,0 +1,5 @@
+export const statuses = {
+  YES: `yes`,
+  NO: `no`
+};
+
